moved transmit pane into transmission store. removed watch interval for close, now looking for close event from board
