+++ plugins/editor/transmit-pane.js
@@ -4,11 +4,10 @@ const React = require('react');
 const { createContainer } = require('sovereign');
 
 const styles = require('./styles');
-const deviceStore = require('../../src/stores/device');
-const { transmitInput } = require('../../src/actions/device');
+const transmissionStore = require('../../src/stores/transmission');
+const { transmitInput } = require('../../src/actions/transmission');
 
 class TransmitPane extends React.Component {
-
 
   handleKeyDown(event) {
     const { keyCode } = event.nativeEvent;
@@ -25,12 +24,12 @@ class TransmitPane extends React.Component {
   }
 
   render() {
-    const { connection, transmitText } = this.props;
+    const { connected, transmitText } = this.props;
     return (
       <div style={styles.transmit}>
         <textarea style={styles.transmitInput} name='transmitInput'
           value={ transmitText } onKeyDown={this.handleKeyDown}
-          onKeyPress={this.handleKeyPress} rows='1' disabled={!connection} />
+          onKeyPress={this.handleKeyPress} rows='1' disabled={!connected} />
         <br />
       </div>
     );
@@ -40,11 +39,11 @@ class TransmitPane extends React.Component {
 module.exports = createContainer(TransmitPane, {
   getStores(){
     return {
-      deviceStore: deviceStore
+      transmissionStore: transmissionStore
     };
   },
 
   getPropsFromStores() {
-    return deviceStore.getState();
+    return transmissionStore.getState();
   }
 });
