+++ src/actions/device.js
@@ -16,10 +16,6 @@ class DeviceActions {
     this.dispatch();
   }
 
-  transmitInput(input) {
-    this.dispatch(input);
-  }
-
   updateSelected(device) {
     this.dispatch(device);
   }
