+++ src/actions/transmission.js
@@ -3,12 +3,25 @@
 const alt = require('../alt');
 
 class TransmissionActions {
+
+  connected() {
+    this.dispatch();
+  }
+
+  disconnected() {
+    this.dispatch();
+  }
+
   rx() {
     this.dispatch();
   }
 
   tx() {
     this.dispatch();
+  }
+
+  transmitInput(input) {
+    this.dispatch(input);
   }
 }
